feat[api-server]: 테스트 시 적용할 인증 설정 추가
